requirements.txt
aiohttp~=3.8.3
lxml~=4.9.1
PyYaml~=6.0
APScheduler~=3.10.1